Реализовал генерацию номера банковской карты, изменил CardData.

+++ src/main/java/com/example/demo/service/CardService.java
@@ -5,5 +5,6 @@ import com.example.demo.dto.CardDto;
 public interface CardService {
 
     CardDto cardIssue(CardData cardData);
-    CardDto getCardById(Integer id);
+    CardDto getCardById(Long id);
+
 }
